+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyselect (development version)
 
+* `num_range()` gains a `suffix` argument (#229).
+
 * `eval_select()` gains new `allow_empty` argument which allows you to prevent
   empty selections with `allow_empty = FALSE` (#252).